refactor repository check and sync functions
